change prompt extra credit done
I could never get the io stuff to work so this is the first part that
works. The comhand is loaded as a process. We will get atleast a some
credit for it. I also did the change prompt extra credit one. If you
have either of the other two you can add it.